feat: expose inThread on MentionNotificationItem
When someone @-mentions another user inside a thread, the resulting
MentionNotificationItem didn't carry the thread root. The frontend
worked around this by routing notification clicks through the
/m/[messageId] permalink resolver, which fetched the event server-side
and re-routed to the thread URL.

Add inThread to the MentionNotification protobuf and the
MentionNotificationItem GraphQL type, populate it in
notifyMentionedUsers, and surface it in the GraphQL converter. The
frontend now reads it via notificationTarget.threadRootId for all
notification types, so getCleanPath returns the thread URL directly
and the /m/ workaround branch goes away.

Click handlers for the notification dot can now pre-set the pending
highlight unconditionally — the destination is always known up-front,
no resolver round-trip needed.

Pulled the helper functions out of role_permissions.resolvers.go into
a separate role_permissions_helpers.go file so gqlgen's codegen
doesn't quarantine them as 'unknown code' on regen.

RoomMessageNotification doesn't need an inThread field — that
notification type is only ever created for root messages (the
ALL_MESSAGES path is gated on inThread == \"\").
