+++ frontend/src/lib/RoomList.svelte
@@ -406,9 +406,7 @@ rooms are organized into collapsible sections. Otherwise, rooms display alphabet
     }
 
     const target = notificationTarget(notification);
-    // Pre-set pending highlight only when we know the thread context (see
-    // InstanceSpaceSection for the rationale).
-    if (target.threadRootId && target.eventId && target.spaceId && target.roomId) {
+    if (target.eventId && target.spaceId && target.roomId) {
       stores.pendingHighlights.set(target.spaceId, target.roomId, target.threadRootId, target.eventId);
     }
     void notificationStore.dismiss(notification.id);
