+++ frontend/src/lib/gql/gql.ts
@@ -78,7 +78,7 @@ type Documents = {
     "\n\tquery GetActiveCallRoomIds($spaceId: ID!) {\n\t\tactiveCallRoomIds(spaceId: $spaceId)\n\t}\n": typeof types.GetActiveCallRoomIdsDocument,
     "\n\tquery GetSidebarCallParticipants($spaceId: ID!, $roomId: ID!) {\n\t\tcallParticipants(spaceId: $spaceId, roomId: $roomId) {\n\t\t\tuserId\n\t\t\tdisplayName\n\t\t\tlogin\n\t\t\tavatarUrl\n\t\t\tjoinedAt\n\t\t}\n\t}\n": typeof types.GetSidebarCallParticipantsDocument,
     "\n\tquery GetCallParticipants($spaceId: ID!, $roomId: ID!) {\n\t\tcallParticipants(spaceId: $spaceId, roomId: $roomId) {\n\t\t\tuserId\n\t\t\tdisplayName\n\t\t\tlogin\n\t\t\tavatarUrl\n\t\t\tjoinedAt\n\t\t}\n\t}\n": typeof types.GetCallParticipantsDocument,
-    "\n  query Notifications {\n    notifications {\n      __typename\n      ... on DMMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        room {\n          id\n        }\n      }\n      ... on MentionNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        mentionSpace: space {\n          id\n          name\n        }\n        mentionRoom: room {\n          id\n          name\n        }\n        mentionEventId: eventId\n      }\n      ... on ReplyNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        replySpace: space {\n          id\n          name\n        }\n        replyRoom: room {\n          id\n          name\n        }\n        replyEventId: eventId\n        inReplyToId\n        replyInThread: inThread\n      }\n      ... on RoomMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        roomMsgSpace: space {\n          id\n          name\n        }\n        roomMsgRoom: room {\n          id\n          name\n        }\n        roomMsgEventId: eventId\n      }\n    }\n  }\n": typeof types.NotificationsDocument,
+    "\n  query Notifications {\n    notifications {\n      __typename\n      ... on DMMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        room {\n          id\n        }\n      }\n      ... on MentionNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        mentionSpace: space {\n          id\n          name\n        }\n        mentionRoom: room {\n          id\n          name\n        }\n        mentionEventId: eventId\n        mentionInThread: inThread\n      }\n      ... on ReplyNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        replySpace: space {\n          id\n          name\n        }\n        replyRoom: room {\n          id\n          name\n        }\n        replyEventId: eventId\n        inReplyToId\n        replyInThread: inThread\n      }\n      ... on RoomMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        roomMsgSpace: space {\n          id\n          name\n        }\n        roomMsgRoom: room {\n          id\n          name\n        }\n        roomMsgEventId: eventId\n      }\n    }\n  }\n": typeof types.NotificationsDocument,
     "\n  query HasNotifications {\n    hasNotifications\n  }\n": typeof types.HasNotificationsDocument,
     "\n  mutation DismissNotification($input: DismissNotificationInput!) {\n    dismissNotification(input: $input)\n  }\n": typeof types.DismissNotificationDocument,
     "\n  mutation DismissAllNotifications {\n    dismissAllNotifications\n  }\n": typeof types.DismissAllNotificationsDocument,
@@ -232,7 +232,7 @@ const documents: Documents = {
     "\n\tquery GetActiveCallRoomIds($spaceId: ID!) {\n\t\tactiveCallRoomIds(spaceId: $spaceId)\n\t}\n": types.GetActiveCallRoomIdsDocument,
     "\n\tquery GetSidebarCallParticipants($spaceId: ID!, $roomId: ID!) {\n\t\tcallParticipants(spaceId: $spaceId, roomId: $roomId) {\n\t\t\tuserId\n\t\t\tdisplayName\n\t\t\tlogin\n\t\t\tavatarUrl\n\t\t\tjoinedAt\n\t\t}\n\t}\n": types.GetSidebarCallParticipantsDocument,
     "\n\tquery GetCallParticipants($spaceId: ID!, $roomId: ID!) {\n\t\tcallParticipants(spaceId: $spaceId, roomId: $roomId) {\n\t\t\tuserId\n\t\t\tdisplayName\n\t\t\tlogin\n\t\t\tavatarUrl\n\t\t\tjoinedAt\n\t\t}\n\t}\n": types.GetCallParticipantsDocument,
-    "\n  query Notifications {\n    notifications {\n      __typename\n      ... on DMMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        room {\n          id\n        }\n      }\n      ... on MentionNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        mentionSpace: space {\n          id\n          name\n        }\n        mentionRoom: room {\n          id\n          name\n        }\n        mentionEventId: eventId\n      }\n      ... on ReplyNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        replySpace: space {\n          id\n          name\n        }\n        replyRoom: room {\n          id\n          name\n        }\n        replyEventId: eventId\n        inReplyToId\n        replyInThread: inThread\n      }\n      ... on RoomMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        roomMsgSpace: space {\n          id\n          name\n        }\n        roomMsgRoom: room {\n          id\n          name\n        }\n        roomMsgEventId: eventId\n      }\n    }\n  }\n": types.NotificationsDocument,
+    "\n  query Notifications {\n    notifications {\n      __typename\n      ... on DMMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        room {\n          id\n        }\n      }\n      ... on MentionNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        mentionSpace: space {\n          id\n          name\n        }\n        mentionRoom: room {\n          id\n          name\n        }\n        mentionEventId: eventId\n        mentionInThread: inThread\n      }\n      ... on ReplyNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        replySpace: space {\n          id\n          name\n        }\n        replyRoom: room {\n          id\n          name\n        }\n        replyEventId: eventId\n        inReplyToId\n        replyInThread: inThread\n      }\n      ... on RoomMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        roomMsgSpace: space {\n          id\n          name\n        }\n        roomMsgRoom: room {\n          id\n          name\n        }\n        roomMsgEventId: eventId\n      }\n    }\n  }\n": types.NotificationsDocument,
     "\n  query HasNotifications {\n    hasNotifications\n  }\n": types.HasNotificationsDocument,
     "\n  mutation DismissNotification($input: DismissNotificationInput!) {\n    dismissNotification(input: $input)\n  }\n": types.DismissNotificationDocument,
     "\n  mutation DismissAllNotifications {\n    dismissAllNotifications\n  }\n": types.DismissAllNotificationsDocument,
@@ -595,7 +595,7 @@ export function graphql(source: "\n\tquery GetCallParticipants($spaceId: ID!, $r
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query Notifications {\n    notifications {\n      __typename\n      ... on DMMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        room {\n          id\n        }\n      }\n      ... on MentionNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        mentionSpace: space {\n          id\n          name\n        }\n        mentionRoom: room {\n          id\n          name\n        }\n        mentionEventId: eventId\n      }\n      ... on ReplyNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        replySpace: space {\n          id\n          name\n        }\n        replyRoom: room {\n          id\n          name\n        }\n        replyEventId: eventId\n        inReplyToId\n        replyInThread: inThread\n      }\n      ... on RoomMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        roomMsgSpace: space {\n          id\n          name\n        }\n        roomMsgRoom: room {\n          id\n          name\n        }\n        roomMsgEventId: eventId\n      }\n    }\n  }\n"): (typeof documents)["\n  query Notifications {\n    notifications {\n      __typename\n      ... on DMMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        room {\n          id\n        }\n      }\n      ... on MentionNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        mentionSpace: space {\n          id\n          name\n        }\n        mentionRoom: room {\n          id\n          name\n        }\n        mentionEventId: eventId\n      }\n      ... on ReplyNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        replySpace: space {\n          id\n          name\n        }\n        replyRoom: room {\n          id\n          name\n        }\n        replyEventId: eventId\n        inReplyToId\n        replyInThread: inThread\n      }\n      ... on RoomMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        roomMsgSpace: space {\n          id\n          name\n        }\n        roomMsgRoom: room {\n          id\n          name\n        }\n        roomMsgEventId: eventId\n      }\n    }\n  }\n"];
+export function graphql(source: "\n  query Notifications {\n    notifications {\n      __typename\n      ... on DMMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        room {\n          id\n        }\n      }\n      ... on MentionNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        mentionSpace: space {\n          id\n          name\n        }\n        mentionRoom: room {\n          id\n          name\n        }\n        mentionEventId: eventId\n        mentionInThread: inThread\n      }\n      ... on ReplyNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        replySpace: space {\n          id\n          name\n        }\n        replyRoom: room {\n          id\n          name\n        }\n        replyEventId: eventId\n        inReplyToId\n        replyInThread: inThread\n      }\n      ... on RoomMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        roomMsgSpace: space {\n          id\n          name\n        }\n        roomMsgRoom: room {\n          id\n          name\n        }\n        roomMsgEventId: eventId\n      }\n    }\n  }\n"): (typeof documents)["\n  query Notifications {\n    notifications {\n      __typename\n      ... on DMMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        room {\n          id\n        }\n      }\n      ... on MentionNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        mentionSpace: space {\n          id\n          name\n        }\n        mentionRoom: room {\n          id\n          name\n        }\n        mentionEventId: eventId\n        mentionInThread: inThread\n      }\n      ... on ReplyNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        replySpace: space {\n          id\n          name\n        }\n        replyRoom: room {\n          id\n          name\n        }\n        replyEventId: eventId\n        inReplyToId\n        replyInThread: inThread\n      }\n      ... on RoomMessageNotificationItem {\n        id\n        createdAt\n        actor {\n          id\n          login\n          displayName\n          avatarUrl(width: 96, height: 96)\n          presenceStatus\n        }\n        summary\n        roomMsgSpace: space {\n          id\n          name\n        }\n        roomMsgRoom: room {\n          id\n          name\n        }\n        roomMsgEventId: eventId\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
