+++ frontend/src/routes/chat/notifications/+page.svelte
@@ -96,9 +96,7 @@
     const store = stores.notifications;
 
     const target = notificationTarget(item.notification);
-    // Pre-set pending highlight only when we know the thread context (see
-    // InstanceSpaceSection for the rationale).
-    if (target.threadRootId && target.eventId && target.spaceId && target.roomId) {
+    if (target.eventId && target.spaceId && target.roomId) {
       stores.pendingHighlights.set(target.spaceId, target.roomId, target.threadRootId, target.eventId);
     }
     void store.dismiss(item.notification.id);
